@@ -21,20 +21,22 @@
     "@pancakeswap/sdk": "^2.4.0",
     "@reduxjs/toolkit": "^1.4.0",
     "@remix-project/remixd": "0.5.1",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
     "@traderjoe-xyz/sdk": "^0.8.6",
     "@types/humanize-plus": "^1.8.0",
-    "@types/jest": "^26.0.9",
-    "@types/node": "^14.0.27",
-    "@types/react": "^16.9.46",
+    "@types/jest": "^29.2.4",
+    "@types/node": "^18.11.11",
+    "@types/react": "^18.0.26",
     "@types/react-redux": "^7.1.9",
     "@walletconnect/web3-provider": "^1.7.8",
     "@web3-react/core": "^6.1.9",
     "@web3-react/injected-connector": "^6.0.7",
+    "@web3-react/magic-connector": "^6.1.9",
     "@web3-react/walletconnect-connector": "^6.2.13",
     "@web3-react/walletlink-connector": "^6.2.14",
+    "@web3auth/web3auth": "^1.2.0",
     "axios": "^0.21.4",
     "bignumber.js": "^9.0.0",
     "classnames": "^2.3.1",
@@ -44,22 +46,26 @@
     "install": "^0.13.0",
     "material-ui-image": "^3.3.2",
     "moment": "^2.29.1",
+    "moralis": "^1.8.0",
+    "moralis-admin-cli": "^2.1.16",
     "numeral": "^2.0.6",
-    "react": "^16.13.1",
+    "react": "^18.2.0",
     "react-circular-progressbar": "^2.0.3",
     "react-countdown": "^2.2.1",
     "react-countup": "^4.3.3",
     "react-device-detect": "^2.1.2",
-    "react-dom": "^16.13.1",
+    "react-dom": "^18.2.0",
     "react-feather": "^2.0.8",
     "react-input-range": "^1.3.0",
+    "react-intersection-observer": "^9.4.0",
     "react-intl": "^5.15.8",
     "react-loading-skeleton": "^2.1.1",
     "react-meta-tags": "^1.0.1",
     "react-redux": "^7.2.1",
     "react-responsive-carousel": "^3.2.22",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.1",
+    "react-router": "6",
+    "react-router-dom": "6",
+    "react-scripts": "^4.0.3",
     "react-slick": "^0.28.1",
     "react-spinners": "^0.11.0",
     "react-spring": "^8.0.27",
@@ -70,7 +76,7 @@
     "sass": "^1.49.7",
     "slick-carousel": "^1.8.1",
     "styled-components": "^5.1.1",
-    "typescript": "^3.9.7",
+    "typescript": "^4.7.4",
     "use-wallet": "OGBoogieTomb/use-wallet",
     "web-vitals": "^1.0.1",
     "web3": "^1.2.11",
@@ -109,8 +115,8 @@
   },
   "devDependencies": {
     "@types/numeral": "0.0.28",
-    "@types/react-dom": "^16.9.8",
-    "@types/react-router-dom": "^5.1.5",
+    "@types/react-dom": "^18.0.9",
+    "@types/react-router-dom": "^5.3.3",
     "@types/styled-components": "^5.1.2",
     "css-loader": "^4.3.0",
     "prettier": "2.3.1"
