@@ -1,6 +1,10 @@
 import React from 'react';
 
-const Card: React.FC = ({children}) => <div>{children}</div>;
+interface Props {
+    children: React.ReactNode;
+  }
+  
+const Card: React.FC<Props> = ({children}) => <div>{children}</div>;
 
 export default Card;
 
