@@ -1,6 +1,8 @@
 import React from 'react';
 import styled from 'styled-components';
-
+import classNames from "classnames";
+import classes from "classnames";
+import styles from './index.module.scss';
 import Card from '../Card';
 import CardContent from '../CardContent';
 import Container from '../Container';
@@ -9,13 +11,13 @@ export interface ModalProps {
     onDismiss?: () => void;
 }
 
-const Modal: React.FC = ({ children }) => {
+const Modal: React.FC<{ children: React.ReactNode | React.ReactNode[] }> = ({ children }) => {
     return (
-        <Container>
+        <Container size="sm">
             <StyledModal>
-                <Card>
-                    <CardContent>{children}</CardContent>
-                </Card>
+                <div className={styles.wrapper}>
+                    {children}
+                </div>
             </StyledModal>
         </Container>
     );
@@ -23,7 +25,6 @@ const Modal: React.FC = ({ children }) => {
 
 const StyledModal = styled.div`
   border-radius: 12px;
-  position: relative;
 `;
 
 export default Modal;
