@@ -7,7 +7,8 @@ import ModalActions from '../../../components/ModalActions';
 import ModalTitle from '../../../components/ModalTitle';
 import TokenInput from '../../../components/TokenInput';
 import styled from 'styled-components';
-
+import classes from "classnames";
+import styles from './index.module.scss';
 import { getDisplayBalance } from '../../../utils/formatBalance';
 import Label from '../../../components/Label';
 import useLpStats from '../../../hooks/useLpStats';
@@ -99,6 +100,7 @@ const ZapModal: React.FC<ZapProps> = ({ onConfirm, onDismiss, tokenName = '', de
     if (!isNumeric(e.currentTarget.value)) return;
     setVal(e.currentTarget.value);
     if (showEstimates) {
+      console.log(e.currentTarget.value)
       const estimateZap = await tombFinance.estimateZapIn(zappingToken, tokenName, String(e.currentTarget.value));
       setEstimate({ token0: estimateZap[0].toString(), token1: estimateZap[1].toString() });
     }
@@ -112,49 +114,67 @@ const ZapModal: React.FC<ZapProps> = ({ onConfirm, onDismiss, tokenName = '', de
     }
   };
 
-  return (
-    <Modal>
-      <ModalTitle text={`Zap in ${tokenName}`} />
+  const handleClick = (e) => {
+    const action = e.target.getAttribute('data-close')
 
-      <StyledActionSpacer />
-      <InputLabel style={{ color: '#000', marginBottom: '1rem' }} id="label">
-        Select Token
-      </InputLabel>
-      <br />
-      <Select onChange={handleChangeAsset} style={{ border: '1px solid black', borderRadius: '10px', background: 'rgb(8, 9, 13, 1, 0.9)', padding: '10px' }} labelId="label" id="select" value={zappingToken}>
-        <StyledMenuItem value={FTM_TICKER}>DAI</StyledMenuItem>
-        {/* <StyledMenuItem value={AVAX_TICKER}>AVAX</StyledMenuItem> */}
-        <StyledMenuItem value={TSHARE_TICKER}>STRAW</StyledMenuItem>
-        <StyledMenuItem value={TOMB_TICKER}>FUDGE</StyledMenuItem>
-      </Select>
-      <TokenInput
-        onSelectMax={handleSelectMax}
-        onChange={handleChange}
-        value={val}
-        max={zappingTokenBalance}
-        symbol={zappingToken}
-      />
-      <br />
-      {showEstimates && <><Label variant="primary" text="Zap Estimations" />
-        <br />
-        <StyledDescriptionText>
-          {' '}
-          {tokenName}: {Number(estimate.token0) / Number(ftmAmountPerLP)}
-        </StyledDescriptionText>
-        {tokenName.startsWith(TOMB_TICKER) ?
-          <StyledDescriptionText>
-            {' '}
-            ({Number(estimate.token0)} {tokenName.startsWith(TOMB_TICKER) ? FTM_TICKER : TOMB_TICKER} /{' '}
-            {Number(estimate.token1)} {tokenName.startsWith(TOMB_TICKER) ? TOMB_TICKER : FTM_TICKER}){' '}
-          </StyledDescriptionText>
-          :
-          <StyledDescriptionText>
-            {' '}
-            ({Number(estimate.token0)} {tokenName.startsWith(TSHARE_TICKER) ? TSHARE_TICKER : FTM_TICKER} /{' '}
-            {Number(estimate.token1)} {tokenName.startsWith(TSHARE_TICKER) ? FTM_TICKER : TSHARE_TICKER}){' '}
-          </StyledDescriptionText>}
-      </>}
-      <InputLabel style={{ color: '#1d48b6', marginBottom: '1rem', marginTop: '1rem' }} id="label">
+    if (action) {
+      onDismiss()
+    }
+  }
+
+  return (
+    <div
+      className={classes(styles.block && styles.active)}
+      data-close={true}
+      onClick={(event) => {
+        handleClick(event)
+      }}
+    >
+      <div className={classes('gradient-background', styles.content)}>
+        <h5 className={styles.title}>{`Zap in ${tokenName}`}</h5>
+
+        <div className={classes(styles.content)}>
+
+
+          <StyledActionSpacer />
+          <InputLabel style={{ color: '#000', marginBottom: '1rem' }} id="label">
+            Select Token
+          </InputLabel>
+          <br />
+          <Select onChange={handleChangeAsset} style={{ border: '1px solid black', borderRadius: '10px', background: 'rgb(8, 9, 13, 1, 0.9)', padding: '10px' }} labelId="label" id="select" value={zappingToken}>
+            <StyledMenuItem value={FTM_TICKER}>DAI</StyledMenuItem>
+            {/* <StyledMenuItem value={AVAX_TICKER}>AVAX</StyledMenuItem> */}
+            <StyledMenuItem value={TSHARE_TICKER}>STRAW</StyledMenuItem>
+            <StyledMenuItem value={TOMB_TICKER}>FUDGE</StyledMenuItem>
+          </Select>
+          <TokenInput
+            onSelectMax={handleSelectMax}
+            onChange={handleChange}
+            value={val}
+            max={zappingTokenBalance}
+            symbol={zappingToken}
+          />
+          <br />
+          {showEstimates && <><Label variant="primary" text="Zap Estimations" />
+            <br />
+            <StyledDescriptionText>
+              {' '}
+              {tokenName}: {Number(estimate.token0) / Number(ftmAmountPerLP)}
+            </StyledDescriptionText>
+            {tokenName.startsWith(TOMB_TICKER) ?
+              <StyledDescriptionText>
+                {' '}
+                ({Number(estimate.token0)} {tokenName.startsWith(TOMB_TICKER) ? FTM_TICKER : TOMB_TICKER} /{' '}
+                {Number(estimate.token1)} {tokenName.startsWith(TOMB_TICKER) ? TOMB_TICKER : FTM_TICKER}){' '}
+              </StyledDescriptionText>
+              :
+              <StyledDescriptionText>
+                {' '}
+                ({Number(estimate.token0)} {tokenName.startsWith(TSHARE_TICKER) ? TSHARE_TICKER : FTM_TICKER} /{' '}
+                {Number(estimate.token1)} {tokenName.startsWith(TSHARE_TICKER) ? FTM_TICKER : TSHARE_TICKER}){' '}
+              </StyledDescriptionText>}
+          </>}
+          {/* <InputLabel style={{ color: '#1d48b6', marginBottom: '1rem', marginTop: '1rem' }} id="label">
         Slippage Tolerance
       </InputLabel>
       <Input
@@ -166,37 +186,39 @@ const ZapModal: React.FC<ZapProps> = ({ onConfirm, onDismiss, tokenName = '', de
         endAdornment={<div style={{ padding: '1px' }}>%</div>}
         fullWidth={false}
         style={{ maxWidth: '3rem', marginLeft: '14px', border: '1px solid black', borderRadius: '10px', padding: '10px' }}
-      />
-      <ModalActions>
-        <Button
-          color="primary"
-          variant="contained"
-          onClick={() =>
-            approveZapperStatus !== ApprovalState.APPROVED
-              ? approveZapper()
-              : onConfirm(zappingToken, tokenName, val, String(+slippage * 100))
-          }
-        >
-          {approveZapperStatus !== ApprovalState.APPROVED ? 'Approve' : "Zap"}
-        </Button>
-      </ModalActions>
+      /> */}
+          <ModalActions>
+            <Button
+              color="primary"
+              variant="contained"
+              onClick={() =>
+                approveZapperStatus !== ApprovalState.APPROVED
+                  ? approveZapper()
+                  : onConfirm(zappingToken, tokenName, val, String(+slippage * 100))
+              }
+            >
+              {approveZapperStatus !== ApprovalState.APPROVED ? 'Approve' : "Zap"}
+            </Button>
+          </ModalActions>
 
-      {/* <StyledActionSpacer />
+          {/* <StyledActionSpacer />
       <Alert variant="outlined" severity="info">
         New feature. Use at your own risk!
       </Alert> */}
-    </Modal>
+        </div>
+      </div>
+    </div>
   );
 };
 
 const StyledActionSpacer = styled.div`
-  height: ${(props) => props.theme.spacing[4]}px;
-  width: ${(props) => props.theme.spacing[4]}px;
+  height: 20px;
+  width: 20px;
 `;
 
 const StyledDescriptionText = styled.div`
   align-items: center;
-  color: ${(props) => props.theme.color.grey[400]};
+  color: grey;
   display: flex;
   font-size: 14px;
   font-weight: 700;
