@@ -2,7 +2,7 @@
 //import { Fetcher as FetcherSpirit, Token as TokenSpirit } from '@traderjoe-xyz/sdk';
 import { Fetcher, Route, Token } from '@traderjoe-xyz/sdk';
 import { Configuration } from './config';
-import { ContractName, TokenStat, AllocationTime, LPStat, Bank, PoolStats, TShareSwapperStat } from './types';
+import { ContractName, TokenStat, AllocationTime, LPStat, Bank, PoolStats, TShareSwapperStat, PegPoolToken,PegPool,PegPoolUserInfo} from './types';
 import { BigNumber, Contract, ethers, Event, EventFilter } from 'ethers';
 import { decimalToBalance } from './ether-utils';
 import { TransactionResponse } from '@ethersproject/providers';
@@ -12,7 +12,8 @@ import { getDefaultProvider } from '../utils/provider';
 import IUniswapV2PairABI from './IUniswapV2Pair.abi.json';
 import config, { bankDefinitions } from '../config';
 import moment from 'moment';
-import { parseUnits } from 'ethers/lib/utils';
+import { commify, formatEther } from 'ethers/lib/utils';
+import { formatUnits, parseUnits } from 'ethers/lib/utils';
 import { FTM_TICKER, SPOOKY_ROUTER_ADDR, TOMB_TICKER, AVAX_TICKER, TSHARE_TICKER } from '../utils/constants';
 /**
  * An API module of 2omb Finance contracts.
@@ -155,7 +156,99 @@ export class TombFinance {
     : await pool.compound();
   }
 
+// // Peg Pool
 
+// async getPegPool(): Promise<PegPool> {
+//   const contract = this.contracts.PegPool;
+//   const busd = new ERC20('0xA7D7079b0FEaD91F3e65f86E8915Cb59c1a4C664', this.signer, 'USDC', 6);
+//   const [depositsEnabled, totalDepositTokenAmount, userInfo, approval] = await Promise.all([
+//     contract.depositsEnabled(),
+//     contract.totalDepositTokenAmount(),
+//     this.getPegPoolUserInfo(),
+//     busd.allowance(this.myAccount, contract.address),
+//   ]);
+
+//   return {
+//     depositsEnabled,
+//     totalDesposits: Number(formatUnits(totalDepositTokenAmount, 6)).toFixed(2),
+//     depositTokenName: 'USDC',
+//     depositToken: busd,
+//     userInfo,
+//     approved: approval.gt(0),
+//   };
+// }
+
+// async getPegPoolUserInfo(): Promise<PegPoolUserInfo> {
+//   const amount: BigNumber = await this.contracts.PegPool.userInfo(this.myAccount);
+//   return {
+//     amountDeposited: getDisplayBalance(amount, 6),
+//     isDeposited: amount.gt(0),
+//     amountDepositedBN: amount,
+//   };
+// }
+
+// async getPegPoolPendingRewards(): Promise<PegPoolToken[]> {
+//   const tokenMap: {
+//     [key: string]: {
+//       name: string;
+//       pair: string;
+//       injection: number;
+//     };
+//   } = {
+//     '0xe6d1aFea0B76C8f51024683DD27FA446dDAF34B6': {
+//       name: 'WSHARE',
+//       pair: '0x03d15E0451e54Eec95ac5AcB5B0a7ce69638c62A',
+//       injection: 0,
+//     },
+//     '0xB31f66AA3C1e785363F0875A1B74E27b85FD66c7': {
+//       name: 'WAVAX',
+//       pair: '0xA389f9430876455C36478DeEa9769B7Ca4E3DDB1',
+//       injection: 0,
+//     },
+//   };
+
+//   const [tks, tokens] = await Promise.all([
+//     this.contracts.PegPool.getRewardTokens(),
+//     this.contracts.PegPool.pendingRewards(this.myAccount),
+//   ]);
+//   const addresses = tokens[0];
+//   const amounts = tokens[1];
+//   const rewards: PegPoolToken[] = [];
+
+//   for (let i = 0; i < addresses.length; i++) {
+//     const info = tokenMap[addresses[i]];
+//     rewards.push({
+//       token: new ERC20(addresses[i], this.provider.getSigner(), info.name),
+//       name: info.name,
+//       pairAddress: info.pair,
+//       amount: Number(formatUnits(amounts[i])).toFixed(8),
+//       pendingValueBN: amounts[i],
+//       rewardPerBlock: Number(formatEther(tks[i].rewardPerBlock)),
+//       canCompound: info.name != 'AALTO',
+//     });
+//   }
+//   return rewards;
+// }
+
+// async depositPegPool(amount: BigNumber) {
+//   return this.contracts.PegPool.deposit(amount);
+// }
+
+// async compoundRewardsPegPool() {
+//   return this.contracts.PegPool.compound();
+// }
+
+// async compoundTokenPegPool() {
+//   return this.contracts.PegPool.compound();
+// }
+
+// async withdrawPegPool(amount: BigNumber) {
+//   return this.contracts.PegPool.withdraw(amount);
+// }
+
+// async claimPegPool() {
+//   return this.contracts.PegPool.claim();
+// }
 
 
   //===================================================================
@@ -682,9 +775,12 @@ export class TombFinance {
 
     const TSHAREPrice = (await this.getShareStat()).priceInDollars;
     const masonrytShareBalanceOf = await this.TSHARE.balanceOf(this.currentMasonry().address);
+    const TBONDPrice = (await this.getBondStat()).priceInDollars;
+    const TBONDTotal = (await this.getBondStat()).totalSupply;
+    const bondTVL = Number(TBONDPrice) * Number(TBONDTotal);
     const masonryTVL = Number(getDisplayBalance(masonrytShareBalanceOf, this.TSHARE.decimal)) * Number(TSHAREPrice);
 
-    return totalValue + masonryTVL;
+    return totalValue + masonryTVL + bondTVL;
   }
 
   /**
